% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object of the relationship of a response
variable and explanatory variables in the columns of \code{data}.}

\item{data}{Input data frame containing a response variable and
explanatory variables.}
}
\value{
A table given by the lm() function.
}
\description{
This function fits a linear model according to a function and a data frame,
and returns a summary table.
}
\examples{
my_lm(lifeExp ~ gdpPercap + continent, data = my_gapminder)

}
\keyword{prediction}
