% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation function}
\usage{
my_rf_cv(trian, cl, k)
}
\arguments{
\item{cl}{True values of the training data \code{train}.}

\item{k}{A numeric indicating the number of folds.}

\item{train}{A given data frame.}
}
\value{
A numeric indicating the mean square error.
}
\description{
This function performs a \code{k} fold random forest cross-validation
on the penguin data.
}
\examples{
my_penguins <- na.omit(my_penguins)
penguins <- my_penguins[,3:5]
my_rf_cv(penguins, my_penguins$body_mass_g, 2)
my_rf_cv(penguins, my_penguins$body_mass_g, 5)
my_rf_cv(penguins, my_penguins$body_mass_g, 10)

}
\keyword{prediction}
