% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric set of data.}

\item{alternative}{A string indicating the t-test type for \code{x} and
\code{mu}, with input choices "less", "greater", and "two.sided".}

\item{mu}{Numeric value indicating the null hypothesis value of the mean.}
}
\value{
A list containing a numeric indicating the test statistic, degree
of freedom, a string indicating the type of t-test, and the p-value of
the t-test.
}
\description{
This function performs a sample t-test based on the given input.
}
\examples{
my_t.test(my_gapminder[[4]], "less", 1)
my_t.test(my_gapminder[[4]], "greater", 3)
my_t.test(my_gapminder[[4]], "two.sided", 6)

}
\keyword{prediction}
